requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
anthropic==0.39.0
qdrant-client==1.12.1
openai==1.54.5
python-dotenv==1.0.1
pydantic==2.10.3

